+++ cmd/neofs-node/config/policer/policer.go
@@ -10,8 +10,6 @@ const (
 	ReplicationCooldownDefault = 1 * time.Second
 	// ObjectBatchSizeDefault is the default replication's objects batch size.
 	ObjectBatchSizeDefault = 10
-	// MaxWorkersDefault is the default replication's worker pool's maximum size.
-	MaxWorkersDefault = 20
 	// BoostMultiplierDefault is the default multiplier for object batch size when
 	// the policer detects missing replicas and performs recovery.
 	BoostMultiplierDefault = 3
@@ -22,7 +20,6 @@ type Policer struct {
 	HeadTimeout         time.Duration `mapstructure:"head_timeout"`
 	ReplicationCooldown time.Duration `mapstructure:"replication_cooldown"`
 	ObjectBatchSize     uint32        `mapstructure:"object_batch_size"`
-	MaxWorkers          uint32        `mapstructure:"max_workers"`
 	BoostMultiplier     uint32        `mapstructure:"boost_multiplier"`
 }
 
@@ -38,9 +35,6 @@ func (p *Policer) Normalize() {
 	}
 	if p.ObjectBatchSize <= 0 {
 		p.ObjectBatchSize = ObjectBatchSizeDefault
-	}
-	if p.MaxWorkers <= 0 {
-		p.MaxWorkers = MaxWorkersDefault
 	}
 	if p.BoostMultiplier <= 0 {
 		p.BoostMultiplier = BoostMultiplierDefault